autorizacion crud projects

+++ uptask-backend/app/Http/Controllers/API/ProjectsController.php
@@ -2,40 +2,61 @@
 
 namespace App\Http\Controllers\API;
 
+use App\Models\Project;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
+use App\Http\Resources\ProjectCollection;
 use App\Http\Requests\Project\StoreRequest;
 use App\Http\Requests\Project\UpdateRequest;
-use App\Http\Resources\ProjectCollection;
-use App\Models\Project;
 
 class ProjectsController extends Controller
 {
     
     public function index()
     {
-        return new ProjectCollection(Project::all());
+        $projects = auth()->user()->projects;
+        return new ProjectCollection($projects);
     }
 
     public function store(StoreRequest $request)
     {
-        $data = $request->validated();
-        return response()->json(['data' => Project::create($data), 'message' => 'Created project succesful'], 201);
+        $project = new Project($request->validated());
+        $result = auth()->user()->projects()->save($project);
+        return response()->json(['data' => $result, 'message' => 'Created project succesful'], 201);
     }
 
     public function show(string $projectId)
     {
         $project = Project::find($projectId);
+
+        if(!Gate::allows('show', $project)) {
+            return response()->json(['error' => 'Unauthorized access'], 403);
+        }
+
         return response()->json(['project' => $project, 'tasks' => $project->tasks]);
     }
 
     public function update(UpdateRequest $request, string $projectId)
     {
+        $project = Project::find($projectId);
+
+        if(!Gate::allows('update', $project)) {
+            return response()->json(['error' => 'Unauthorized access'], 403);
+        }
+
         $project = Project::where('id', $projectId)->update($request->validated());
+
         return response()->json(['data' => $project, 'message' => 'Updated project succesful']);
     }
 
     public function destroy(string $projectId)
     {
+        $project = Project::find($projectId);
+
+        if(!Gate::allows('delete', $project)) {
+            return response()->json(['error' => 'Unauthorized access'], 403);
+        }
+
         $project = Project::where('id', $projectId)->delete();
         return response()->json(['data' => $project, 'message' => 'Deleted project succesful']);
     }
